requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.4
django-autocomplete-light==3.8.2
django-ckeditor==6.1.0
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
django-redis==5.0.0
djangorestframework==3.12.4
Fabric3==1.14.post1
idna==2.10
invoke==1.5.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mistune==0.8.4
mysqlclient==2.0.3
paramiko==2.7.2
pathlib2==2.3.5
pycparser==2.20
PyNaCl==1.4.0
pypiserver==1.4.2
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
